Relocate cloudformation function
