+++ internal/cloudformation/BUILD
@@ -1,7 +1,7 @@
 subinclude("///pleasings2//go:compat")
 
 go_library(
-    name = "distribution",
+    name = "cloudformation",
     srcs = glob(
         ["*.go"],
         exclude = ["*_test.go"],
